refactor: 사주 패널 중복 코드 통합 (918줄 → 667줄)
학생/선생님 사주 분석 패널의 90% 중복 코드를 공통 컴포넌트로 추출:
- SajuPanel: 공통 UI/상태/로직 (503줄)
- SajuAnalysisPanel: 학생용 래퍼 (81줄, 히스토리 포함)
- TeacherSajuPanel: 선생님용 래퍼 (83줄)
- 차이점은 콜백 props로 주입 (action, simplify, reset)
